+++ demo-app/src/different-vars-on-rerender/changing-component.js
@@ -1,11 +1,12 @@
-import { customElement, html, css } from 'lit-element';
+import { customElement, LitElement, html, css } from 'lit-element';
+import { observeState } from '@app/lit-state.js';
 import { DemoComponent } from '@app/demo-component.js';
 import '@app/components/code-small';
 import { demoState } from './state';
 
 
 @customElement('changing-component')
-export class ChangingComponent extends DemoComponent {
+export class ChangingComponent extends DemoComponent(observeState(LitElement)) {
 
     render() {