+++ demo-app/src/different-vars-on-rerender/index.js
@@ -75,10 +75,9 @@ export class DifferentVarsOnRerender extends DemoElement {
                     You don't have to worry about which
                     <code-small>stateVar</code-small> or
                     <code-small>asyncStateVar</code-small> you render at which
-                    time. As long as your component extends from
-                    <code-small>LitStateElement</code-small> or uses the mixin
-                    <code-small><a href="#mixin-usage">LitStateElementMixin</a></code-small>,
-                    LitState will keep your component synchronized.
+                    time. As long as your component uses the
+                    <code-small>observeState</code-small> mixin, your component
+                    will stay synchronized.
                 </p>
 
             </div>
@@ -89,13 +88,13 @@ export class DifferentVarsOnRerender extends DemoElement {
 
     get changingComponentCode() {
 
-        return `import { customElement, html } from 'lit-element';
-import { LitStateElement } from 'lit-element-state';
+        return `import { customElement, LitElement, html } from 'lit-element';
+import { observeState } from 'lit-element-state';
 import { demoState } from './demo-state.js';
 
 
 @customElement('changing-component')
-export class ChangingComponent extends LitStateElement {
+export class ChangingComponent extends observeState(LitElement) {
 
     render() {