+++ demo-app/src/mixin-usage/state.js
@@ -1,9 +1,0 @@
-import { LitState, stateVar } from '../lit-state.js';
-
-
-class DemoState extends LitState {
-    counter = stateVar(0);
-}
-
-
-export const demoState = new DemoState();