+++ demo-app/src/state-var/index.js
@@ -41,9 +41,8 @@ export class StateVar extends DemoElement {
                 </p>
 
                 <p>
-                    The components that use the state extend from
-                    <code-small>LitStateElement</code-small> instead of
-                    <code-small>LitElement</code-small>. This makes them
+                    The components that use the state use the mixin
+                    <code-small>observeState</code-small>. This makes them
                     automatically re-render when a
                     <code-small>stateVar</code-small> they use changes:
                 </p>
@@ -82,12 +81,12 @@ export const demoState = new DemoState();`;
 
     get componentCode() {
 
-        return `import { customElement, html } from 'lit-element';
-import { LitStateElement } from 'lit-element-state';
+        return `import { customElement, LitElement, html } from 'lit-element';
+import { observeState } from 'lit-element-state';
 import { demoState } from './demo-state.js';
 
 @customElement('component-1')
-export class Component1 extends LitStateElement {
+export class Component1 extends observeState(LitElement) {
 
     render() {
         return html\`