+++ demo-app/src/state-var/state-var-component-1.js
@@ -1,10 +1,11 @@
-import { customElement, html, css } from 'lit-element';
+import { customElement, LitElement, html, css } from 'lit-element';
+import { observeState } from '@app/lit-state.js';
 import { DemoComponent } from '@app/demo-component.js';
 import { demoState } from './state';
 
 
 @customElement('state-var-component-1')
-export class StateVarComponent1 extends DemoComponent {
+export class StateVarComponent1 extends DemoComponent(observeState(LitElement)) {
 
     render() {
         return html`